% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTVEM_boot.R
\name{CTVEM_boot}
\alias{CTVEM_boot}
\title{Continuous Time-Varying Effect model for a bootstrapping estimation.}
\usage{
CTVEM_boot(
  data = NULL,
  Time = "Time",
  outcome = NULL,
  ID = "ID",
  estimate = "marginal",
  iterations = 50,
  quantiles = c(0.025, 0.975),
  boot = TRUE,
  pivot = "Mean",
  Tpred = seq(0, 30, 1),
  plot_show = FALSE,
  ncores = NULL,
  standardized = TRUE,
  method = "bam",
  gamma = 1,
  k = 10,
  ktrend = 3,
  ctype = "PSOCK"
)
}
\arguments{
\item{data}{Specify the data frame that contains the interested data, Time (measuing time) and ID column. MUST INCLUDE COLNAMES.}

\item{Time}{The name of the Time variable. E.G. Time = "Time" (must be specified).}

\item{outcome}{This is the outcome variables. Specified as outcome="outcomevariablename" for a single variable or outcome=c("outcomevariablename1","outcomevariablename2"). If it is NULL, it will consider each variables as outcome once.}

\item{ID}{The name of the ID column in the data E.G. ID = "ID"}

\item{estimate}{The relationship which we are interested, estimate = "marginal" or "partial". Default is the "marginal".}

\item{iterations}{How many times bootstrapping single estimation you want to perform, the default value is 50}

\item{quantiles}{The quantiles to build bootstrapping CI, the default value is c(low_quantile, high_quantile) = c(.025, 0.975)}

\item{boot}{Indicate if we perform bootstrapping estimation or not. If boot == True, we perfomr bootstrapping estimation. The default value is False}

\item{pivot}{Indicate the pivot of the bootstrapping CI is the mean of bootstrapping point estimations or the median of bootstrapping point estimations. The default value is ''Mean''}

\item{Tpred}{A vector which indicates that interested time points, e.g. seq(0,30,1)}

\item{plot_show}{The option to suppress the plot outcomes. The default if FALSE which means the plot outcomes will not appear.}

\item{ncores}{How many cores you want to use. If it is null, ncores = detectCores()/2}

\item{standardized}{This specifies whether all of the variables (aside from Time) should be standardized. Options are TRUE, FALSE, and "center". TRUE means within-person standardize each variable (aka get the person-centered z-scores), FALSE means use the raw data, "center" means to only within-person mean-center the variables. Default = TRUE. FALSE is not recommended unless you have done these transformations yourself (OPTIONAL)}

\item{method}{Indicate which method will be used to estimate time-varying effetcs. The default value is "bam". Another option is "gam".}

\item{gamma}{This can be used to change the wiggliness of the model. This can be useful if the model is too smooth (i.e flat). The lower the number the more wiggly this will be (see ?gam in MGCV for more information). The default is equal to 1. (OPTIONAL, UNCOMMONLY SPECIFIED)}

\item{k}{The number of k selection points used in the model for stage 1 (see ?choose.k in mgcv package for more details) The ideal k is the maximum number of data points per person, but this slows down DTVEM and is often not required. (OPTIONAL, BUT RECOMMENDED)}

\item{ktrend}{The number of k selection points used in the model for the time spline (NOTE THAT THIS CONTROLS FOR TIME TRENDS OF THE POPULATION)  (see ?choose.k in mgcv package for more details). Default is 3. (OPTIONAL)}
}
\value{
a) point estimates, b) highCI, c) lowCI, but now based on the bootstrap_results
}
\description{
This is the CTVEM function which performs bootstrapping estimation. The output of this function is the point estimations, high and low cooresponding CIs.
}
