% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTVEM_single.R
\name{CTVEM_single}
\alias{CTVEM_single}
\title{Continuous Time-Varying Effect model for a single-time bootstrapping estimation.}
\usage{
CTVEM_single(
  data = NULL,
  Time = "Time",
  outcome = NULL,
  ID = "ID",
  estimate = "marginal",
  Tpred = seq(0, 30, 1),
  plot_show = FALSE,
  boot = FALSE,
  output_type = "CI",
  standardized = TRUE,
  method = "bam",
  gamma = 1,
  k = 10,
  ktrend = 3,
  quantiles = c(0.025, 0.975),
  predictionsend = NULL,
  weighting = FALSE
)
}
\arguments{
\item{data}{Specify the data frame that contains the interested data, Time (measuing time) and ID column. MUST INCLUDE COLNAMES.}

\item{Time}{The name of the Time column in the data E.G. Time = "Time" (must be specified).}

\item{outcome}{This is the outcome variables. Specified as outcome="outcomevariablename" for a single variable or outcome=c("outcomevariablename1","outcomevariablename2"). If it is NULL, it will consider each variables as outcome once.}

\item{ID}{The name of the ID column in the data E.G. ID = "ID"}

\item{estimate}{The relationship which we are interested, estimate = "marginal" or "partial". Default is the "marginal".}

\item{Tpred}{A vector which indicates that interested time points, e.g. seq(0,30,1)}

\item{plot_show}{The option to suppress the plot outcomes. The default if FALSE which means the plot outcomes will not appear.}

\item{boot}{Indicate if we perform bootstrapping estimation or not. If boot == True, we perfomr bootstrapping estimation. The default value is False}

\item{output_type}{Indicate which output form will be returned. If output_type == "CI", point estimations and corresponding CIs will be returned. If output_type =="PE", only ponit estimation will be returned. If output_type =="SCI", the Simultaneous CIs will be returned. The default value is "CI"}

\item{standardized}{This specifies whether all of the variables (aside from Time) should be standardized. Options are TRUE, FALSE, and "center". TRUE means within-person standardize each variable (aka get the person-centered z-scores), FALSE means use the raw data, "center" means to only within-person mean-center the variables. Default = TRUE. FALSE is not recommended unless you have done these transformations yourself (OPTIONAL)}

\item{method}{Indicate which method will be used to estimate time-varying effetcs. The default value is "bam". Another option is "gam".}

\item{gamma}{This can be used to change the wiggliness of the model. This can be useful if the model is too smooth (i.e flat). The lower the number the more wiggly this will be (see ?gam in MGCV for more information). The default is equal to 1. (OPTIONAL, UNCOMMONLY SPECIFIED)}

\item{k}{The number of k selection points used in the model for stage 1 (see ?choose.k in mgcv package for more details) The ideal k is the maximum number of data points per person, but this slows down DTVEM and is often not required. (OPTIONAL, BUT RECOMMENDED)}

\item{ktrend}{The number of k selection points used in the model for the time spline (NOTE THAT THIS CONTROLS FOR TIME TRENDS OF THE POPULATION)  (see ?choose.k in mgcv package for more details). Default is 3. (OPTIONAL)}

\item{quantiles}{The quantiles to build bootstrapping CI, the default value is c(low_quantile, high_quantile) = c(.025, 0.975)}

\item{weighting}{Test option. Calculate weights for stacked data vector?}
}
\description{
This is the CTVEM function which serves for performing bootstrapping estimation. The output of this function is the single estimation of the bootstrapping process, i.e., randomly select data rows from the original dataset, then do estimation.
}
