% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManipulation.R
\name{datamanipulation_old}
\alias{datamanipulation_old}
\title{Data manipulation}
\usage{
datamanipulation_old(
  input_list = input_list,
  differentialtimevaryingpredictors = NULL,
  outcome = NULL,
  data = NULL,
  ID = "ID",
  Time = NULL,
  standardized = TRUE,
  predictionstart = NULL,
  predictionsend = NULL
)
}
\arguments{
\item{input_list}{identical to the elipses above, A list of variable names used in the function e.g. "X","Y"}

\item{differentialtimevaryingpredictors}{A list of predictor variables to be used as predictor variables e.g. c("X","Y")}

\item{outcome}{The outcome(s) to be predicted e.g. c("X","Y")}

\item{data}{The data frame to be converted}

\item{ID}{The name of the ID variable. E.G. ID = "ID" (must be specified).}

\item{Time}{The name of the Time variable. E.G. Time = "Time" (must be specified).}

\item{standardized}{This specifies whether all of the variables (aside from Time) should be standardized. Options are TRUE, FALSE, and "center". TRUE means within-person standardize each variable, FALSE means use the raw data, "center" means to only within-person center the variables. Default = TRUE}

\item{predictionstart}{The differential time value to start with, default is NULL, and the lowest time difference in the time series will be used (use lower value if you're first value if you're interested in a smaller interval prediction)}

\item{predictionsend}{The differential time value to end with. This should usually be set by the user.}
}
\value{
The output of this will be (1) The data in a very long stacked format (called laglongreducedummy), (2) the data in a wide format with all lags for all variables, time, and time differences (called Timelagsdummy), (3) the name of the differentialtimevarying predictors (namesofnewpredictorvariables), and (4) the length of the long matrix (laglongmatrixlength)
}
\description{
PLEASE USE THE LAG FUNCTION RATHER THAN THIS UNLESS YOU WOULD LIKE TO SPECIFY THIS MANUALLY. Manipulates the data so that it creates a vector that is both in wide and long formats
}
